refactor: 패키지 구조 개선
- ActorCommand는 공통적으로 사용하는 마커 인터페이스이므로 domain/chat/common으로 이동
- 포트 역할을 수행하는 Protocol의 위치를 domain/chat/port/out으로 이동
