--- pekko/src/main/java/com/tok/pekko/domain/chat/command/ActorCommand.java
@@ -1,4 +1,0 @@
-package com.tok.pekko.domain.chat.command;
-
-public interface ActorCommand {
-}
