+++ pekko/src/main/java/com/tok/pekko/domain/chat/common/ActorCommand.java
@@ -1,0 +1,4 @@
+package com.tok.pekko.domain.chat.common;
+
+public interface ActorCommand {
+}
