+++ pekko/src/main/java/com/tok/pekko/infrastructure/actor/CborSerializable.java
@@ -1,6 +1,6 @@
 package com.tok.pekko.infrastructure.actor;
 
-import com.tok.pekko.domain.chat.command.ActorCommand;
+import com.tok.pekko.domain.chat.common.ActorCommand;
 
 public interface CborSerializable extends ActorCommand {
 }
